requirements.txt
pyzmq>=24.0.0
numpy>=1.23.3
opencv-python>=4.6.0
imageZMQ
depthai>=2.17.3
